requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
contourpy==1.1.1
cycler==0.12.1
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.10.0
idna==3.4
imageai==3.0.3
iniconfig==2.0.0
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mock==4.0.3
mpmath==1.3.0
networkx==3.2
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
py==1.11.0
pyparsing==3.1.1
pytest==7.1.3
python-dateutil==2.8.2
requests==2.31.0
scipy==1.11.3
six==1.16.0
sympy==1.12
tomli==2.0.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.64.1
triton==2.1.0
typing_extensions==4.8.0
urllib3==2.0.7